+++ packages/stream_chat_persistence/test/src/dao/pinned_message_dao_test.dart
@@ -353,11 +353,11 @@ void main() {
     const cid = 'test:Cid';
     const limit = 15;
     const lessThan = 'testMessageId${cid}25';
-    const greaterThanOrEqual = 'testMessageId${cid}5';
+    const greaterThan = 'testMessageId${cid}5';
     const pagination = PaginationParams(
       limit: limit,
       lessThan: lessThan,
-      greaterThanOrEqual: greaterThanOrEqual,
+      greaterThan: greaterThan,
     );
 
     // Should be empty initially
@@ -377,7 +377,7 @@ void main() {
       messagePagination: pagination,
     );
     expect(fetchedMessages.length, limit);
-    expect(fetchedMessages.first.id, greaterThanOrEqual);
+    expect(fetchedMessages.first.id, greaterThan);
     expect(fetchedMessages.last.id != lessThan, true);
   });