+++ packages/stream_chat_flutter_core/CHANGELOG.md
@@ -1,3 +1,13 @@
+## Upcoming
+
+🐞 Fixed
+
+- Fixed `MessageListCore` not properly loading and paginating thread replies.
+
+✅ Added
+
+- Added methods for paginating thread replies in `StreamChannel`.
+
 ## 9.15.0
 
 ✅ Added